convert to Double colors
* with colors in a range from 0.0 to 1.0 we can start to manipulate 
  them with simple arithmetic

+++ src/Surface.hs
@@ -3,7 +3,7 @@ module Surface where
 import Rays
 import Vectors
 
-type Color = (Int, Int, Int)
+type Color = (Double, Double, Double)
 
 data Surface = Sphere Point Double Color
